Fix parsing of records that are not PERF_RECORD_SAMPLE
For all records that are not PERF_RECORD_SAMPLE, if sample_id_all is set in the
attributes for perf_event_open(), extract struct sample_id as appropriate. This
change makes the parsing honor the sample_id_all flag instead of always trying
to parse out that data even when it's not present.