Overhaul tracing 'format' file parsing and decoding
The original code was trying to avoid having to parse out type names to get
field width and type information; however, not enough sizing information is
given by the kernel to be able to do this reliably in all cases; therefore,
parsing type names is necessary. It's ugly, but the parser makes a best effort
to figure out what it needs to know. In some cases, fields are skipped. This
could be changed to include the fields as raw bytes, but it seems wasteful
right now. Pointers, structs, unions, and arrays for which the type cannot be
determined are skipped now.