Be more permissive in the case of unknown type names/sizes
Instead of raising an error for an unrecognized type name, try to figure it
out from the size information that's available. When figuring out the type
from size information, don't raise an error for an unrecognized size. Instead,
mark the type for skipping. It can later either be skipped or presented as an
array of raw bytes