Remove the use of mutexes from the read path of decoders
The decoder map is used very often for reading, but rarely for updating. Make
the read path wait free in all cases. For writes, locks are used to prevent
multiple simultaneous writers from stomping on each other. It's not likely that
writers will experience much contention, if any at all.

An alternative implementation removes the use of a mutex altogether and
replaces the atomic store with a CAS operation, retrying if there's a conflict.
This implementation using a lock is simpler