Move decoding of SampleRecord rawdata decoding off of the hotpath
Previously, decoding of rawdata was occuring during processing of ringbuffer
data, but that's not necessary due to SampleRecord.read() making a copy of the
raw data. Defer decoding until later. Originally, I had intended to actually
decode the raw data, but defer the API callback until later, but this is better
given the data copy that happens in SampleRecord.read()