Only monitor groupfds for readiness rather than every event fd
There are only as many groupfds as there are CPUs, but there are n*ncpus event
fds. Due to the way the perf_event interface works, we only need to monitor the
groupfds rather than all of the eventfds. In addition, we don't even need to
read the data from the ready fds; all we need to do is empty the ring buffer
and the next poll will not continue to flag the fd as ready (unless more data
has been added to the ringbuffer in the interim). This allows us to remove more
complexity and system calls during ring buffer processing.