Replace lossy process cache and remove process lineage.
Process lineage tracking is best performed by telemetry
subscribers in order to keep the sensor as close to stateless
as possible. Process lineage is also problematic in that the
sensor can track the kernel's notion of process hierarchy, but
caching that is difficult. Alternatively, tracking the original
parents is not reconstructable after a sensor restart.

The previous process cache had trouble assigning container IDs
to processes that were created before the container running event
has been processed. This resulted in containerIDs not being
assigned properly to events when the kernel_compile benchmark
container was being run.