Fix shadowing redeclarations of `err` and perf reference event.
The PERF_TYPE_HW event we were using wasn't working inside a
VM, so replace it with the DUMMY SW event. DUMMY isn't necessarily
ideal since it requires kernel 3.12, but we also use it elsewhere.
When we want to replace it to support older kernels, we should just
replace both of them.