Fix timing issues with group leader initialization
During the initialization of an EventMonitor instance when the group
leaders for a cgroup/pid are initialized to compute clock offsets,
make sure to clear out the ring buffers after the events used to
compute clock offsets are closed. The sample frequency is set to a
single clock tick, so it's possible (and likely) that stray events
will get put into the queue between the time when the offset is
computed from the first event and the time that the event is closed,
which can then cause problems later, resulting in unrecognized
SampleID error messages in the logs, or worse, crashes.