Update container start event identification for Docker < 1.13.
Versions of Docker prior to 1.13 include a runc that does not
change the process names to the runc:[0:INIT] style names, so
we need to match on a comm value of "exe" instead.

Older versions of Docker on CentOS also use a native cgroup driver,
so the cgroup path is different. Use a regular expression to
identify Docker container IDs from cgroup paths for more robust
identification.