Use a kprobe for syscall enter events rather than a tracepoint
There are two problems with using the tracepoint raw_syscalls/syscall_enter
as we were previously doing. First, various kernel versions have bugs around
how they pass the argument information back to userspace. This was previously
worked around in a less-than-elegant fashion. Second, perf_event filters do
not support indexing into arrays, so it is not possible to filter based on
arguments. Both problems can be solved using a kprobe, but there's a catch:
the kprobe will never fire unless the tracepoint is enabled. The way around
this is to enable the tracepoint with a filter that will never allow it to
add data to the perf_event ringbuffer.