Use a kprobe to track process command-line information
Use a kprobe in the process info cache to track command-line information
when a process execs. This will make the command-line information available
later for short-lived processes instead of relying on procfs.

Due to limitations with the kprobe interface, the number of argv elements
must be fixed, so the information cached may not always be complete.